fix: build step - forkTs point to correct file - move BUILD_SERVER_FILE_SERVER to @shuvi/shared
