+++ packages/shuvi/src/constants.ts
@@ -21,5 +21,3 @@ export const BUILD_CLIENT_RUNTIME_WEBPACK = `static/runtime/webpack.js`;
 export const BUILD_CLIENT_RUNTIME_POLYFILL = `static/runtime/polyfill.js`;
 
 export const BUILD_SERVER_DIR = 'server';
-
-export const BUILD_SERVER_FILE_SERVER = `server.js`;
