+++ packages/toolpack/src/utils/forkTsCheckerWebpackPlugin.ts
@@ -1,5 +1,5 @@
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 export { Issue } from 'fork-ts-checker-webpack-plugin/lib/issue';
-export { createCodeFrameFormatter } from 'fork-ts-checker-webpack-plugin/lib/formatter/codeFrameFormatter';
+export { createCodeFrameFormatter } from 'fork-ts-checker-webpack-plugin/lib/formatter/CodeFrameFormatter';
 
 export default ForkTsCheckerWebpackPlugin;
